Update JS rules and linter
I believe that most of us agree that airbnb's eslint configuration
should be our default linter for JavaScript. I know that, for some
projects, we may feel that tweaking it would be better, but I believe
that we should agree in a set of rules and follow them. We'll never find
something that everyone likes at first. I also think that we need a
default configuration because it will remove this conversation out of
the way, and let us focus on different things. I also think that we need
consistency in our projects, and we need to remove the amount of
decisions we make whenever we start a new one. This doesn't mean that
this topic is closed forever, we should be open to new ideas, but keep
in mind that the benefits of using one styleguide or the other are
probably very small. Most times, I believe, it's a matter of taste.